+++ src/control_data_cache.h
@@ -5,16 +5,16 @@
 #define CACHE_CHANNEL_LEN 21
 
 typedef struct {
-    char  channel[CACHE_CHANNEL_LEN];
-    float value;
-    bool  received;
+    char   channel[CACHE_CHANNEL_LEN];
+    double value;
+    bool   received;
 } control_data_cache_item;
 
 class control_data_cache {
 public:
     void init();
     void update_from_json(char* json); // NOTE: mutates json buffer (tiny-json)
-    bool fetch_value(const char* channel, float& out) const;
+    bool fetch_value(const char* channel, double& out) const;
     bool was_received(const char* channel) const;
     uint8_t count() const { return _count; }
     void clear() { _count = 0; }
